created the separate component for listing editors
